+++ js/codec/BasicCodec.ts
@@ -18,6 +18,7 @@ class AVGCalculator {
     }
 }
 
+declare class webkitOfflineAudioContext extends OfflineAudioContext {}
 abstract class BasicCodec implements Codec {
     protected _audioContext: OfflineAudioContext;
     protected _decodeResampler: AudioResampler;
@@ -32,7 +33,7 @@ abstract class BasicCodec implements Codec {
     constructor(codecSampleRate: number) {
         this.channelCount = 1;
         this.samplesPerUnit = 960;
-        this._audioContext = new OfflineAudioContext(AudioController.globalContext.destination.channelCount, 1024,AudioController.globalContext.sampleRate );
+        this._audioContext = new (webkitOfflineAudioContext || OfflineAudioContext)(AudioController.globalContext.destination.channelCount, 1024,AudioController.globalContext.sampleRate );
         this._codecSampleRate = codecSampleRate;
         this._decodeResampler = new AudioResampler(AudioController.globalContext.sampleRate);
         this._encodeResampler = new AudioResampler(codecSampleRate);