+++ js/voice/AudioResampler.ts
@@ -1,5 +1,6 @@
 class AudioResampler {
     targetSampleRate: number;
+    private _use_promise: boolean;
 
     constructor(targetSampleRate: number = 44100){
         this.targetSampleRate = targetSampleRate;
@@ -7,18 +8,37 @@ class AudioResampler {
     }
 
     resample(buffer: AudioBuffer) : Promise<AudioBuffer> {
+        if(!buffer) {
+            console.warn("Received empty buffer as input! Returning empty output!");
+            return new Promise<AudioBuffer>(resolve => resolve(undefined));
+        }
         //console.log("Encode from %i to %i", buffer.sampleRate, this.targetSampleRate);
         if(buffer.sampleRate == this.targetSampleRate)
             return new Promise<AudioBuffer>(resolve => resolve(buffer));
 
         let context;
-        context = new OfflineAudioContext(buffer.numberOfChannels, Math.ceil(buffer.length * this.targetSampleRate / buffer.sampleRate), this.targetSampleRate);
+        context = new (webkitOfflineAudioContext || OfflineAudioContext)(buffer.numberOfChannels, Math.ceil(buffer.length * this.targetSampleRate / buffer.sampleRate), this.targetSampleRate);
 
         let source = context.createBufferSource();
         source.buffer = buffer;
-        source.connect(context.destination);
         source.start(0);
+        source.connect(context.destination);
 
-        return context.startRendering();
+        if(typeof(this._use_promise) === "undefined") {
+           this._use_promise = navigator.browserSpecs.name != 'Safari';
+        }
+
+        if(this._use_promise)
+            return context.startRendering();
+        else {
+            return new Promise<AudioBuffer>((resolve, reject) => {
+                context.oncomplete = event => resolve(event.renderedBuffer);
+                try {
+                    context.startRendering();
+                } catch (ex) {
+                    reject(ex);
+                }
+            })
+        }
     }
 }